Modulo de detalhes finalizado

+++ app/Http/Controllers/Admin/PlanController.php
@@ -56,10 +56,22 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
-
+        $plan = $this->repository
+                    ->with('details')
+                    ->where('url', $url)->first();
         if(!$plan)
             return redirect()->back();
+
+        /**
+         * Checa se existem detalhes vinculados ao plano
+         */
+
+        if($plan->details->count() > 0){ 
+
+            return redirect()
+                    ->back()
+                    ->with('error', 'Existem detalhes vinculados ao plano, por favor delete-os antes!');
+        }
 
         $plan->delete();
